+++ example/main.dart
@@ -1,86 +1,21 @@
-import 'dart:convert';
-
-import 'dart:ffi';
-import 'package:ffi/ffi.dart';
 import 'package:gpiod/proxy_gpiod.dart';
-import 'package:logging/logging.dart';
-import 'dart:ffi';
-import 'dart:isolate';
-import 'dart:typed_data';
 
-
-final Logger _log = new Logger('main');
-
-
-class CppRequest {
-  final SendPort replyPort;
-  final int pendingCall;
-  final String method;
-  final Uint8List data;
-
-  factory CppRequest.fromCppMessage(List message) {
-    return CppRequest._(message[0], message[1], message[2], message[3]);
-  }
-
-  CppRequest._(this.replyPort, this.pendingCall, this.method, this.data);
-
-  String toString() => 'CppRequest(method: $method, ${data.length} bytes)';
-}
-
-void handleCppRequests(dynamic message) {
-  final cppRequest = CppRequest.fromCppMessage(message);
-  print('Dart:   Got message: $cppRequest');
-
-//  if (cppRequest.method == 'myCallback1') {
-//    // Use the data in any way you like. Here we just take the first byte as
-//    // the argument to the function.
-//    final int argument = cppRequest.data[0];
-//    final int result = myCallback1(argument);
-//    final cppResponse =
-//    CppResponse(cppRequest.pendingCall, Uint8List.fromList([result]));
-//    print('Dart:   Responding: $cppResponse');
-//    cppRequest.replyPort.send(cppResponse.toCppMessage());
-//  } else if (cppRequest.method == 'myCallback2') {
-//    final int argument = cppRequest.data[0];
-//    myCallback2(argument);
-//  }
-}
 
 void main() async {
-  Logger.root.level = Level.ALL;
-  Logger.root.onRecord.listen((LogRecord rec) {
-    StringBuffer message = new StringBuffer();
-    message.write('${rec.level.name}:${rec.loggerName} ${rec.time}: ${rec.message}');
-    if (rec.error != null) {
-      message.write(' ${rec.error}');
-    }
-    print(message);
-  });
-
   try {
-
-//var gpiod = GPIOD();
-//var pthread = Pthread();
-
-    var proxyGPIOD = new ProxyGPIOD();
-    proxyGPIOD.print_hello();
-
-    final interactiveCppRequests = ReceivePort()..listen(handleCppRequests);
-    final int nativePort = interactiveCppRequests.sendPort.nativePort;
-    proxyGPIOD.register_send_port(nativePort);
-    final gpio = await FlutterGpiod.getInstance();
+    final gpio = ProxyGpiod.getInstance();
     _status(gpio.chips);
-    _led(gpio.chips);
-    _pirTest(gpio.chips);
-    _button(gpio.chips);
-
+    await _led(gpio.chips);
+    await _pirTest(gpio.chips);
+    await _button(gpio.chips);
+    _status(gpio.chips);
   } catch(e, st){
-    _log.severe("Error",e, st);
+    print('Error $e $st');
   }
 }
 
 
-Future _pirTest(List<GpioChip> chips) async {
+Future _pirTest(List<GpioChip> chips) async{
   print('PIR');
   final lineLED =
   chips.singleWhere((chip) => chip.label == 'pinctrl-bcm2835').lines[14];
@@ -88,28 +23,28 @@ Future _pirTest(List<GpioChip> chips) async {
   chips.singleWhere((chip) => chip.label == 'pinctrl-bcm2835').lines[17];
 
   /// Request BCM 14 as output.
-  await lineLED.requestOutput(
+  lineLED.requestOutput(
       consumer: "flutter_gpiod test", initialValue: true);
 
   for (var i = 0; i<2 ;i++) {
     /// Pulse the line.
     /// Set it to inactive. (so low voltage = GND)
-    await lineLED.setValue(false);
+    lineLED.setValue(false);
     await Future.delayed(Duration(milliseconds: 500));
-    await lineLED.setValue(true);
+    lineLED.setValue(true);
     await Future.delayed(Duration(milliseconds: 500));
   }
-  await lineLED.setValue(false);
+  lineLED.setValue(false);
 
-  await linePIR.requestInput(
+  linePIR.requestInput(
       consumer: "PIR",
       //        activeState: ActiveState.low,
       triggers: {SignalEdge.falling, SignalEdge.rising});
 
   /// Log line events for eternity.
-  linePIR.onEvent.listen((event) async{
+  linePIR.onEvent.listen((event){
     print("PIR: $event");
-    await lineLED.setValue(event.edge == SignalEdge.falling);
+    lineLED.setValue(event.edge == SignalEdge.falling);
   });
 
 //    await lineLED.release();
@@ -132,26 +67,26 @@ Future _led(List<GpioChip> chips) async {
   chips.singleWhere((chip) => chip.label == 'pinctrl-bcm2835').lines[14];
 
   /// Request BCM 14 as output.
-  await line14.requestOutput(
+  line14.requestOutput(
       consumer: "flutter_gpiod test", initialValue: true);
 
   for (var i = 0; i<5 ;i++) {
     /// Pulse the line.
     /// Set it to inactive. (so low voltage = GND)
-    await line14.setValue(false);
+    line14.setValue(false);
     await Future.delayed(Duration(milliseconds: 500));
-    await line14.setValue(true);
+    line14.setValue(true);
     await Future.delayed(Duration(milliseconds: 500));
   }
-  await line14.release();
+  line14.release();
 }
 
-Future _button(List<GpioChip> chips) async {
+Future _button(List<GpioChip> chips) {
   print('BUTTON');
   final line15 =
   chips.singleWhere((chip) => chip.label == 'pinctrl-bcm2835').lines[15];
 
-  await line15.requestInput(
+  line15.requestInput(
       consumer: "BUTTON",
       activeState: ActiveState.high,
       triggers: {SignalEdge.falling, SignalEdge.rising});
@@ -165,7 +100,7 @@ Future _button(List<GpioChip> chips) async {
 //    await line15.release();
 }
 
-Future _status(List<GpioChip> chips) async {
+Future _status(List<GpioChip> chips) {
   /// Print out all GPIO chips and all lines
   /// for all GPIO chips.
   for (var chip in chips) {
@@ -173,9 +108,9 @@ Future _status(List<GpioChip> chips) async {
 
     var index = 0;
     for (var line in chip.lines) {
-      final info = await line.info;
+      final info = line.info;
       if (info.consumer!=null) {
-        print("$index:  ${await line.info}");
+        print("$index:  ${line.info}");
       }
 
       index++;