+++ lib/src/bindings/proxy_gpiod/bindings.dart
@@ -4,21 +4,49 @@ import 'dart:ffi';
 import 'package:gpiod/src/bindings/proxy_gpiod/signatures.dart';
 import 'package:gpiod/src/bindings/proxy_gpiod/types.dart';
 
-final DynamicLibrary _dynamicLibrary = DynamicLibrary.open(ProxyGPIOD.libraryName);
+final DynamicLibrary _dynamicLibrary = DynamicLibrary.open(LibraryProxyGPIOD.libraryName);
 
-class ProxyGPIOD {
+class LibraryProxyGPIOD {
   static const String libraryName = './libproxy_gpiod.so';
 
 
   //int epoll_ctl(int epfd, int op, int fd, struct epoll_event *event)
 //  int Function(int epfd, int op, int fd, Pointer<epoll_event> event) epoll_ctl;
 
-  ///void print_hello();
-  final void Function() print_hello = _dynamicLibrary.lookup<NativeFunction<print_hello_native_t>>("print_hello")
-      .asFunction();
-
   final register_send_port = _dynamicLibrary.lookupFunction<Void Function(Int64 sendPort),
       void Function(int sendPort)>('proxy_gpiod_register_send_port');
+  final _registerDart_PostCObject = _dynamicLibrary.lookupFunction<
+      Void Function(
+          Pointer<NativeFunction<Int8 Function(Int64, Pointer<Dart_CObject>)>>
+          functionPointer),
+      void Function(
+          Pointer<NativeFunction<Int8 Function(Int64, Pointer<Dart_CObject>)>>
+          functionPointer)>('proxy_gpiod_register_dart_postcobject');
+
+  final _registerDart_NewNativePort = _dynamicLibrary.lookupFunction<
+      Void Function(
+          Pointer<
+              NativeFunction<
+                  Int64 Function(
+                      Pointer<Uint8>,
+                      Pointer<NativeFunction<Dart_NativeMessageHandler>>,
+                      Int8)>>
+          functionPointer),
+      void Function(
+          Pointer<
+              NativeFunction<
+                  Int64 Function(
+                      Pointer<Uint8>,
+                      Pointer<NativeFunction<Dart_NativeMessageHandler>>,
+                      Int8)>>
+          functionPointer)>('proxy_gpiod_register_dart_newnativeport');
+
+  final _registerDart_CloseNativePort = _dynamicLibrary.lookupFunction<
+      Void Function(
+          Pointer<NativeFunction<Int8 Function(Int64)>> functionPointer),
+      void Function(
+          Pointer<NativeFunction<Int8 Function(Int64)>> functionPointer)>(
+      'proxy_gpiod_register_dart_closenativeport');
 
   ///int gpiodp_get_num_chips(struct platch_obj *object, FlutterPlatformMessageResponseHandle *responsehandle)
   final int Function() get_num_chips = _dynamicLibrary.lookup<NativeFunction<gpiodp_get_num_chips_native_t>>('gpiodp_get_num_chips')
@@ -37,14 +65,17 @@ class ProxyGPIOD {
   final int Function() supports_reconfiguration = _dynamicLibrary.lookup<NativeFunction<gpiodp_get_num_chips_native_t>>('gpiodp_supports_reconfiguration')
       .asFunction();
 
-  static ProxyGPIOD _instance;
+  static LibraryProxyGPIOD _instance;
 
-  factory ProxyGPIOD(){
+  factory LibraryProxyGPIOD(){
     if (_instance == null) {
-      _instance = new ProxyGPIOD._internal();
+      _instance = new LibraryProxyGPIOD._internal();
+      _instance._registerDart_PostCObject(NativeApi.postCObject);
+      _instance._registerDart_NewNativePort(NativeApi.newNativePort);
+      _instance._registerDart_CloseNativePort(NativeApi.closeNativePort);
     }
     return _instance;
   }
 
-  ProxyGPIOD._internal();
+  LibraryProxyGPIOD._internal();
 }